% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmLooStats.R
\name{kmLooStats}
\alias{kmLooStats}
\title{Simple Leave-one-out statistics for a km (DiceKriging) model}
\usage{
kmLooStats(km, type = "UK")
}
\arguments{
\item{km}{model fit as a km (DiceKriging) object}

\item{type}{Typr of kriging, one of 'SK' (simple kriging), or 'UK' (universal kriging)}
}
\value{
A list of statistics summarising leave-one-out prediction performance. loo is the prediction, mae is Mean Absolute Error,
maxerr is the maximum error and pmae is the mean absolute error expressed as a proportion of the range of the output.
}
\description{
Simple Leave-one-out statistics for a km (DiceKriging) model
}
