% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize columns of a matrix onto a \link{0-1} basis}
\usage{
normalize(X, wrt = NULL)
}
\arguments{
\item{X}{A design matrix}

\item{wrt}{A design matrix which acts as a reference to normalize X against}
}
\value{
A matrix normalized to \link{0-1}
}
\description{
Normalize columns of a matrix onto a \link{0-1} basis
}
