% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaat_design.R
\name{oaat_design}
\alias{oaat_design}
\title{Create a one-at-a-time perturbation design.}
\usage{
oaat_design(X, n, med = TRUE, hold = NULL)
}
\arguments{
\item{X}{Design matrix, e.g. latin hypercube.}

\item{n}{Number of required design points in each dimension.}

\item{med}{Use median rather than mean?}

\item{hold}{Points at which to hold inputs, rather than mean or median.}
}
\value{
A matrix of design points.
}
\description{
Create a one-at-a-time perturbation design.
}
\examples{
oaat_design(X, 21)
}
