% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_beta.R
\name{samp_beta}
\alias{samp_beta}
\title{A matrix of samples from beta distributions}
\usage{
samp_beta(n, mins, maxes, wrt, shape1 = 2, shape2 = 2)
}
\arguments{
\item{n}{Desired number of samples.}

\item{mins}{vector containing minima of samples in original space.}

\item{maxes}{vector containing maxima of sample in original space.}

\item{wrt}{a design that samples should be normalized to.}

\item{shape1}{shape parameter for the beta distribution.}

\item{shape2}{shape parameter for the beta distribution.}
}
\value{
list containing
\itemize{
\item x.unn - matrix of beta distributions in the space defined by mins, maxes
\item x.nwrt - matrix of beta distributions normalized to the space desribed by wrt
}
}
\description{
A matrix of samples from beta distributions
}
