% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_norm.R
\name{samp_norm}
\alias{samp_norm}
\title{A matrix of samples from normal distributions}
\usage{
samp_norm(n, means, sds)
}
\arguments{
\item{n}{number of desired samples.}

\item{means}{vector of desired means, one per column.}

\item{sds}{vector of desired standard deviations, one per column.}
}
\value{
matrix containing samples from normal distributions in the columns.
}
\description{
A matrix of samples from normal distributions
}
\examples{
samp_unif(20, means = c(1,10), sds = c(1,10))
}
