% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_unif.R
\name{samp_unif}
\alias{samp_unif}
\title{A matrix of samples from a uniform distribution}
\usage{
samp_unif(n, mins, maxes)
}
\arguments{
\item{n}{number of rows}

\item{mins}{desired column minima}

\item{maxes}{desirde column maxima}
}
\value{
A matrix of samples from a uniform distribution
}
\description{
A matrix of samples from a uniform distribution
}
\examples{
samp_unif(20, mins = c(0,0), maxes = c(1,1))
}
