% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taat_design.R
\name{taat_design}
\alias{taat_design}
\title{Two-at-a-time (taat) parameter perturbation design}
\usage{
taat_design(X, n, means = NULL)
}
\arguments{
\item{X}{Design matrix on which to base the taat design.}

\item{n}{number of samples per dimension.}

\item{means}{values at which to hold the non-perturbed parameters.}
}
\value{
matrix containing two-at-a-time perturbation design. Each block of rows perturbs two parameters, with all other parameters held at mean, or supplied values.
}
\description{
Two-at-a-time (taat) parameter perturbation design
}
\examples{
taat_design(X, n = 21)
}
