% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnormalize.R
\name{unnormalize}
\alias{unnormalize}
\title{unnormalize a matrix}
\usage{
unnormalize(X_n, un_mins, un_maxes)
}
\arguments{
\item{X_n}{matrix on \link{0-1} basis}

\item{un_mins}{vecror of mins of unnormalized matrix, elements corresponding to columns of X_n}

\item{un_maxes}{}
}
\value{
unnormalized matrix
}
\description{
unnormalize returns a matrix normalized to the \link{0-1} basis back to its unnormalized state,
given reference vectors of mins and maxes.
}
\examples{
unnormalize(X_n, un_mins, un_maxes)
}
